requirements.txt
python-dotenv==0.21.0
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==7.1.2
exceptiongroup==1.0.4
Flask==1.1.2
gunicorn==20.0.4
h11==0.14.0
idna==3.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
outcome==1.2.0
PyJWT==2.6.0
PySocks==1.7.1
pytz==2022.6
requests==2.28.1
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
twilio==7.15.4
urllib3==1.26.13
Werkzeug==1.0.1
wsproto==1.2.0